Move utilities and combinators to a separate package
Hyper works with all combinators removed, as would anything that depends
on the initial 0.4 release.

Removing methods from the trait is a breaking change, but it is possible
to do the semver trick (release 0.5, then release 0.4.x with a
dependency on the 0.5 trait) to not split the ecosystem.

Addresses #52.
