+++ .github/workflows/CI.yml
@@ -27,4 +27,4 @@ jobs:
       - name: Install Rust
         run: rustup update ${{ matrix.rust }} && rustup default ${{ matrix.rust }}
       - name: Run tests
-        run: cargo test
+        run: cargo test --workspace
