+++ Cargo.toml
@@ -1,34 +1,3 @@
-[package]
-name = "http-body"
-# When releasing to crates.io:
-# - Remove path dependencies
-# - Update html_root_url.
-# - Update doc url
-#   - Cargo.toml
-#   - README.md
-# - Update CHANGELOG.md.
-# - Create "vx.y.z" git tag.
-version = "0.4.5"
-authors = [
-  "Carl Lerche <me@carllerche.com>",
-  "Lucio Franco <luciofranco14@gmail.com>",
-  "Sean McArthur <sean@seanmonstar.com>",
-]
-edition = "2018"
-readme = "README.md"
-documentation = "https://docs.rs/http-body"
-repository = "https://github.com/hyperium/http-body"
-license = "MIT"
-description = """
-Trait representing an asynchronous, streaming, HTTP request or response body.
-"""
-keywords = ["http"]
-categories = ["web-programming"]
-
-[dependencies]
-bytes = "1"
-http = "0.2"
-pin-project-lite = "0.2"
-
-[dev-dependencies]
-tokio = { version = "1", features = ["macros", "rt"] }
+[workspace]
+members = ["http-body", "http-body-util"]
+resolver = "2"
