+++ http-body-util/Cargo.toml
@@ -1,0 +1,35 @@
+[package]
+name = "http-body-util"
+# When releasing to crates.io:
+# - Remove path dependencies
+# - Update html_root_url.
+# - Update doc url
+#   - Cargo.toml
+#   - README.md
+# - Update CHANGELOG.md.
+# - Create "http-body-util-x.y.z" git tag.
+version = "0.1.0"
+authors = [
+  "Carl Lerche <me@carllerche.com>",
+  "Lucio Franco <luciofranco14@gmail.com>",
+  "Sean McArthur <sean@seanmonstar.com>",
+]
+edition = "2018"
+readme = "README.md"
+documentation = "https://docs.rs/http-body-util"
+repository = "https://github.com/hyperium/http-body"
+license = "MIT"
+description = """
+Combinators and adapters for HTTP request or response bodies.
+"""
+keywords = ["http"]
+categories = ["web-programming"]
+
+[dependencies]
+bytes = "1"
+http = "0.2"
+http-body = { path = "../http-body" }
+pin-project-lite = "0.2"
+
+[dev-dependencies]
+tokio = { version = "1", features = ["macros", "rt"] }
