+++ http-body-util/src/combinators/map_data.rs
@@ -1,4 +1,5 @@
-use crate::Body;
+use bytes::Buf;
+use http_body::Body;
 use pin_project_lite::pin_project;
 use std::{
     any::type_name,
@@ -8,18 +9,18 @@ use std::{
 };
 
 pin_project! {
-    /// Body returned by the [`map_err`] combinator.
+    /// Body returned by the [`map_data`] combinator.
     ///
-    /// [`map_err`]: crate::util::BodyExt::map_err
+    /// [`map_data`]: crate::util::BodyExt::map_data
     #[derive(Clone, Copy)]
-    pub struct MapErr<B, F> {
+    pub struct MapData<B, F> {
         #[pin]
         inner: B,
         f: F
     }
 }
 
-impl<B, F> MapErr<B, F> {
+impl<B, F> MapData<B, F> {
     #[inline]
     pub(crate) fn new(body: B, f: F) -> Self {
         Self { inner: body, f }
@@ -46,13 +47,14 @@ impl<B, F> MapErr<B, F> {
     }
 }
 
-impl<B, F, E> Body for MapErr<B, F>
+impl<B, F, B2> Body for MapData<B, F>
 where
     B: Body,
-    F: FnMut(B::Error) -> E,
+    F: FnMut(B::Data) -> B2,
+    B2: Buf,
 {
-    type Data = B::Data;
-    type Error = E;
+    type Data = B2;
+    type Error = B::Error;
 
     fn poll_data(
         self: Pin<&mut Self>,
@@ -62,8 +64,8 @@ where
         match this.inner.poll_data(cx) {
             Poll::Pending => Poll::Pending,
             Poll::Ready(None) => Poll::Ready(None),
-            Poll::Ready(Some(Ok(data))) => Poll::Ready(Some(Ok(data))),
-            Poll::Ready(Some(Err(err))) => Poll::Ready(Some(Err((this.f)(err)))),
+            Poll::Ready(Some(Ok(data))) => Poll::Ready(Some(Ok((this.f)(data)))),
+            Poll::Ready(Some(Err(err))) => Poll::Ready(Some(Err(err))),
         }
     }
 
@@ -71,25 +73,20 @@ where
         self: Pin<&mut Self>,
         cx: &mut Context<'_>,
     ) -> Poll<Result<Option<http::HeaderMap>, Self::Error>> {
-        let this = self.project();
-        this.inner.poll_trailers(cx).map_err(this.f)
+        self.project().inner.poll_trailers(cx)
     }
 
     fn is_end_stream(&self) -> bool {
         self.inner.is_end_stream()
     }
-
-    fn size_hint(&self) -> crate::SizeHint {
-        self.inner.size_hint()
-    }
 }
 
-impl<B, F> fmt::Debug for MapErr<B, F>
+impl<B, F> fmt::Debug for MapData<B, F>
 where
     B: fmt::Debug,
 {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        f.debug_struct("MapErr")
+        f.debug_struct("MapData")
             .field("inner", &self.inner)
             .field("f", &type_name::<F>())
             .finish()
