+++ http-body/tests/is_end_stream.rs
